refactor: use package-based imports for cross-package references
Moves the DenoJSON type from the tools package to a separate file and exports it from the package's entry point. This allows other packages to import the type using a package-based import instead of a relative path.

This change also adds the @manypkg/tools package as a devDependency to the @manypkg/cli package and updates the import statement in the corresponding test file.
