+++ packages/cli/package.json
@@ -30,6 +30,7 @@
   },
   "devDependencies": {
     "@changesets/types": "^6.1.0",
+    "@manypkg/tools": "workspace:^",
     "@types/validate-npm-package-name": "^4.0.2",
     "fixturez": "^1.1.0",
     "strip-ansi": "^7.1.0",
