+++ packages/tools/src/index.ts
@@ -7,3 +7,4 @@ export { PnpmTool } from "./PnpmTool.ts";
 export { RootTool } from "./RootTool.ts";
 export { RushTool } from "./RushTool.ts";
 export { YarnTool } from "./YarnTool.ts";
+export type { DenoJSON } from "./DenoTool.ts";
